refactor: 금액 관련 타입을 BigDecimal에서 Long으로 수정

+++ auth-service/src/main/java/com/msa/auth/dto/BalanceResponse.java
@@ -1,11 +1,9 @@
 package com.msa.auth.dto;
 
-import java.math.BigDecimal;
-
 /**
  * 예치금 조회 응답 DTO.
  *
  * @param userId  사용자 ID
  * @param balance 현재 예치금 잔액 (원화)
  */
-public record BalanceResponse(Long userId, BigDecimal balance) {}
+public record BalanceResponse(Long userId, Long balance) {}
