+++ auth-service/src/main/java/com/msa/auth/dto/ChargeRequest.java
@@ -3,8 +3,6 @@ package com.msa.auth.dto;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
-import java.math.BigDecimal;
-
 /**
  * 예치금 충전 요청 DTO.
  *
@@ -13,6 +11,6 @@ import java.math.BigDecimal;
 public record ChargeRequest(
         @NotNull(message = "충전 금액은 필수입니다.")
         @Positive(message = "충전 금액은 0보다 커야 합니다.")
-        BigDecimal amount
+        Long amount
 ) {
 }
