+++ common-module/src/main/java/com/msa/common/event/StockReservedEvent.java
@@ -1,6 +1,5 @@
 package com.msa.common.event;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -20,6 +19,6 @@ public record StockReservedEvent(
     Long userId,
     Long productId,
     int quantity,
-    BigDecimal totalAmount
+    Long totalAmount
 ) {
 }
